server: split endpoints.go into batches.go and files.go
I've always had confusion between endpoints.go vs routing.go. This
also brings the code layout to more similarity with Moov's other projects.
